Abstract hook serializers into their own module
